moved settings panel logic to its own file
